requirements.txt
# === Core UI ===
streamlit>=1.37.0

# === Env ===
python-dotenv>=1.0.1
typing_extensions>=4.12.2

# === OpenAI & Agents SDK ===
openai>=1.88.0
openai-agents>=0.2.9
# (If your pip can’t find `openai-agents`, try: openai_agents>=0.2.9)

# === MCP server (for your custom tools) ===
mcp>=1.2.0
fastmcp>=2.1.1
httpx>=0.27.2

# === LangChain (modern split) for RAG, loaders, chains ===
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0

# === LangGraph (for agentic graphs / Excel Viz agents etc.) ===
langgraph>=0.2.0

# === Vector store for RAG ===
faiss-cpu>=1.12.0

# === YouTube transcripts (for YouTube RAG) ===
youtube-transcript-api

# === (Optional, commonly used with CSV/plotting; safe to keep) ===
pandas>=2.2.2